HasGqlType
