+++ src/Element/Concerns/HasGqlType.php
@@ -1,0 +1,50 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CraftCms\Cms\Element\Concerns;
+
+use craft\gql\interfaces\Element as ElementGqlType;
+use GraphQL\Type\Definition\Type;
+use ReflectionClass;
+
+/**
+ * HasGraphQlType provides GraphQL type definition functionality.
+ *
+ * This trait contains methods for defining GraphQL types and scopes
+ * for elements exposed through the GraphQL API.
+ *
+ * @internal
+ */
+trait HasGqlType
+{
+    /**
+     * {@inheritdoc}
+     */
+    public static function baseGqlType(): Type
+    {
+        return ElementGqlType::getType();
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @since 3.3.0
+     */
+    public static function gqlScopesByContext(mixed $context): array
+    {
+        // Default to no scopes required
+        return [];
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @since 3.3.0
+     */
+    public function getGqlTypeName(): string
+    {
+        // Default to the short class name
+        return new ReflectionClass($this)->getShortName();
+    }
+}
